Deploy ArtiFact-Tag v0.1
Deploy ArtiFact-Tag v0.1
v0.1 of the Tag has been deployed for initial experimentation. 

The objective of these first iterations is to find the right balance between comprehensiveness and smooth user experience whilst factoring in scalability and flexibility for the long term.

First of all we want to make sure the information about content being labeled is comprehensive enough to communicate to users and machine how the latter has been involved in the creation process.

Second, we also respect the fact that content creators don't want to go through a convoluted process in order to tag the content they are creating, therefore the need of having a smooth experience for the labeling process.

Third, we need to keep in mind that as the use of AI in content creation grows, it's likely that users will scale the amount of content produced in a day or a week. ArtiFact will need to support scalability of content production so that the content creator will still be motivated in the utilisation of the system.

Ultimately, we need to factor in Flexibility as well. Similarly, the ways that AI is used to create content are likely to evolve over time, we may need to update the tags accordingly.